refactor component name to match naming convention
