+++ wls-gui-wahllokalsystem/src/components/wahlvorbereitung/BaseCardWahleroeffnungSave.vue
